requirements.txt
django
djangorestframework
django-unfold
pymysql
mysqlclient
pillow
django-auto-logout